+++ integration/network/bridge/bridge_linux_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/docker/docker/daemon/libnetwork/nlwrap"
 	ctr "github.com/docker/docker/integration/internal/container"
 	"github.com/docker/docker/integration/internal/network"
-	"github.com/docker/docker/internal/testutils/networking"
+	"github.com/docker/docker/integration/internal/testutils/networking"
 	"github.com/docker/docker/testutil"
 	"github.com/docker/docker/testutil/daemon"
 	"github.com/docker/go-connections/nat"
